refactor: ♻️ extract push.rs from sync_engine module (#191)
